thingsboard/requirements.txt
git+https://github.com/ozgur/python-firebase
matplotlib
numpy
paho-mqtt
pyzmq
requests
numpy
